+++ src/main/java/cn/nukkit/blockentity/BlockEntityFurnace.java
@@ -262,13 +262,13 @@ public class BlockEntityFurnace extends BlockEntitySpawnable implements Inventor
             int windowId = player.getWindowId(this.getInventory());
             if (windowId > 0) {
                 ContainerSetDataPacket pk = new ContainerSetDataPacket();
-                pk.windowid = (byte) windowId;
+                pk.windowid = windowId;
                 pk.property = 0;
                 pk.value = cookTime;
                 player.dataPacket(pk);
 
                 pk = new ContainerSetDataPacket();
-                pk.windowid = (byte) windowId;
+                pk.windowid = windowId;
                 pk.property = 1;
                 pk.value = burnDuration;
                 player.dataPacket(pk);
